feat: Добавлен Webpack, код разбит на модули
1. Добавлен Webpack c модулями webpack-cli и webpack-dev-server
2. Настрен сервер для автоматического обновления
3. Функции отображения вынесены в модули в папке src/view
4. Поправлена адресация файлов в index.html
5. Поправлен .gitignore, добавлен .eslintignore
