+++ src/view/board.js
@@ -1,0 +1,6 @@
+export const makeBoardTemplate = () => {
+  return `<section class="board container">
+        <div class="board__tasks">
+        </div>
+      </section>`;
+};
