+++ src/view/load-more-button.js
@@ -1,0 +1,3 @@
+export const makeLoadMoreButtonTemplate = () => {
+  return `<button class="load-more" type="button">load more</button>`;
+};
